code cleanup: use classes for timeseries, sections in preparation of new api
