--- config/mock_responses/adoption.json
@@ -1,35 +1,0 @@
-{
-  "results": [
-    {
-      "technology_name": "Wordpress",
-      "dates": [
-        {
-          "date": "2022-04-22T00:00:00.000Z",
-          "origins": {
-            "mobile": 524362,
-            "desktop": 73542
-          }
-        },
-        {
-          "date": "2022-03-22T00:00:00.000Z",
-          "origins": {
-            "mobile": 471531,
-            "desktop": 68634
-          }
-        }
-      ]
-    },
-    {
-      "technology_name": "Squarespace",
-      "dates": [
-        {
-          "date": "2022-04-22T00:00:00.000Z",
-          "origins": {
-            "mobile": 16345,
-            "desktop": 735
-          }
-        }
-      ]
-    }
-  ]
-}
