--- config/mock_responses/all.json
@@ -1,57 +1,0 @@
-{
-  "technology": "all",
-  "name": "All Technologies",
-  "metadata": {
-    "description": "This is the data we have across all technologies, from the origins that match the filters. For now this is what we use as fallback if the technology entered isn't found or if no tech is selected."
-  },
-  "dates": [
-    {
-      "date": "2022-04-22T00:00:00.000Z",
-      "tested": 204743,
-      "good_overall": {
-        "desktop": 6354453,
-        "mobile": 6443446
-      },
-      "median_cls": {
-        "desktop": "1",
-        "mobile": "2.4"
-      },
-      "median_fid": {
-        "desktop": "32",
-        "mobile": "29"
-      },
-      "median_lcp": {
-        "desktop": "3",
-        "mobile": "2"
-      },
-      "origins": {
-        "mobile": 91735451,
-        "desktop": 82635351
-      }
-    },
-    {
-      "date": "2022-03-22T00:00:00.000Z",
-      "tested": 204743,
-      "good_overall": {
-        "desktop": 6354453,
-        "mobile": 6443446
-      },
-      "median_cls": {
-        "desktop": "1",
-        "mobile": "2.4"
-      },
-      "median_fid": {
-        "desktop": "32",
-        "mobile": "29"
-      },
-      "median_lcp": {
-        "desktop": "3",
-        "mobile": "2"
-      },
-      "origins": {
-        "mobile": 91735451,
-        "desktop": 82635351
-      }
-    }
-  ]
-}
