--- config/mock_responses/drupal.json
@@ -1,59 +1,0 @@
-{
-  "technology": "drupal",
-  "name": "Drupal",
-  "metadata": {
-    "categories": ["CMS", "Blog"],
-    "description": "Drupal is a free and open-source web content management system written in PHP and distributed under the GNU General Public License.",
-    "similar": ["squarespace", "wordpress"]
-  },
-  "dates": [
-    {
-      "date": "2022-04-22T00:00:00.000Z",
-      "tested": 204743,
-      "good_overall": {
-        "desktop": 6353,
-        "mobile": 746
-      },
-      "median_cls": {
-        "desktop": "0",
-        "mobile": "0"
-      },
-      "median_fid": {
-        "desktop": "24",
-        "mobile": "12"
-      },
-      "median_lcp": {
-        "desktop": "2",
-        "mobile": "1"
-      },
-      "origins": {
-        "mobile": 1635,
-        "desktop": 8363
-      }
-    },
-    {
-      "date": "2022-03-22T00:00:00.000Z",
-      "tested": 104312,
-      "good_overall": {
-        "desktop": 14,
-        "mobile": 263
-      },
-      "median_cls": {
-        "desktop": "0",
-        "mobile": "0"
-      },
-      "median_fid": {
-        "desktop": "24",
-        "mobile": "12"
-      },
-      "median_lcp": {
-        "desktop": "2",
-        "mobile": "1"
-      },
-      "origins": {
-        "mobile": 222,
-        "desktop": 7353
-      }
-    }
-  ]
-}
