--- config/mock_responses/good_vitals.json
@@ -1,42 +1,0 @@
-{
-  "results": [
-    {
-      "technology": "wordpress",
-      "dates": [
-        {
-          "date": "2022-04-22T00:00:00.000Z",
-          "tested": 204743,
-          "good_overall": {
-            "desktop": 73624,
-            "mobile": 13947,
-            "dataset": 937263
-          }
-        },
-        {
-          "date": "2022-03-22T00:00:00.000Z",
-          "tested": 104312,
-          "good_overall": {
-            "desktop": 14,
-            "mobile": 263,
-            "dataset": 424
-          }
-        }
-      ]
-    },
-    {
-      "technology": "squarespace",
-      "dates": [
-        {
-          "date": "2022-04-22T00:00:00.000Z",
-          "name": "overall",
-          "tested": 264,
-          "good_overall": {
-            "desktop": 163,
-            "mobile": 13,
-            "dataset": 54
-          }
-        }
-      ]
-    }
-  ]
-}
