--- config/mock_responses/metadata.json
@@ -1,14 +1,0 @@
-{
-  "results": [
-    {
-      "technology": "wordpress",
-      "title": "Wordpress",
-      "description": "Wordpress description"
-    },
-    {
-      "technology": "squarespace",
-      "title": "Squarespace",
-      "description": "Squarespace description"
-    }
-  ]
-}
