--- config/mock_responses/squarespace.json
@@ -1,59 +1,0 @@
-{
-  "technology": "squarespace",
-  "name": "Squarespace",
-  "metadata": {
-    "categories": ["CMS", "Blog"],
-    "description": "Squarespace, Inc. is an American website building and hosting company which is based in New York City. It provides software as a service for website building and hosting, and allows users to use pre-built website templates and drag-and-drop elements to create and modify webpages",
-    "similar": ["wordpress", "drupal"]
-  },
-  "dates": [
-    {
-      "date": "2022-04-22T00:00:00.000Z",
-      "tested": 204743,
-      "good_overall": {
-        "desktop": 654,
-        "mobile": 837
-      },
-      "median_cls": {
-        "desktop": "0",
-        "mobile": "0"
-      },
-      "median_fid": {
-        "desktop": "24",
-        "mobile": "12"
-      },
-      "median_lcp": {
-        "desktop": "2",
-        "mobile": "1"
-      },
-      "origins": {
-        "mobile": 2735,
-        "desktop": 363
-      }
-    },
-    {
-      "date": "2022-03-22T00:00:00.000Z",
-      "tested": 104312,
-      "good_overall": {
-        "desktop": 14,
-        "mobile": 263
-      },
-      "median_cls": {
-        "desktop": "0",
-        "mobile": "0"
-      },
-      "median_fid": {
-        "desktop": "24",
-        "mobile": "12"
-      },
-      "median_lcp": {
-        "desktop": "2",
-        "mobile": "1"
-      },
-      "origins": {
-        "mobile": 1111,
-        "desktop": 73542
-      }
-    }
-  ]
-}
