--- config/mock_responses/wordpress.json
@@ -1,59 +1,0 @@
-{
-  "technology": "wordpress",
-  "name": "Wordpress",
-  "metadata": {
-    "categories": ["CMS", "Blog"],
-    "description": "WordPress is a web content management system. It was originally created as a tool to publish blogs but has evolved to support publishing other web content, including more traditional websites, mailing lists and Internet forum, media galleries, membership sites, learning management systems and online stores.",
-    "similar": ["squarespace", "drupal"]
-  },
-  "dates": [
-    {
-      "date": "2022-04-22T00:00:00.000Z",
-      "tested": 204743,
-      "good_overall": {
-        "desktop": 73624,
-        "mobile": 13947
-      },
-      "median_cls": {
-        "desktop": "0",
-        "mobile": "0"
-      },
-      "median_fid": {
-        "desktop": "24",
-        "mobile": "12"
-      },
-      "median_lcp": {
-        "desktop": "2",
-        "mobile": "1"
-      },
-      "origins": {
-        "mobile": 524362,
-        "desktop": 73542
-      }
-    },
-    {
-      "date": "2022-03-22T00:00:00.000Z",
-      "tested": 104312,
-      "good_overall": {
-        "desktop": 14,
-        "mobile": 263
-      },
-      "median_cls": {
-        "desktop": "0",
-        "mobile": "0"
-      },
-      "median_fid": {
-        "desktop": "24",
-        "mobile": "12"
-      },
-      "median_lcp": {
-        "desktop": "2",
-        "mobile": "1"
-      },
-      "origins": {
-        "mobile": 524362,
-        "desktop": 73542
-      }
-    }
-  ]
-}
